+++ libs/linux-mpu9150/eMPL/inv_mpu_dmp_motion_driver.h
@@ -61,8 +61,8 @@ int dmp_enable_feature(unsigned short mask);
 int dmp_get_enabled_features(unsigned short *mask);
 int dmp_set_interrupt_mode(unsigned char mode);
 int dmp_set_orientation(unsigned short orient);
-int dmp_set_gyro_bias(long *bias);
-int dmp_set_accel_bias(long *bias);
+int dmp_set_gyro_bias(int32_t *bias);
+int dmp_set_accel_bias(int32_t *bias);
 
 /* Tap functions. */
 int dmp_register_tap_cb(void (*func)(unsigned char, unsigned char));
@@ -71,7 +71,7 @@ int dmp_set_tap_axes(unsigned char axis);
 int dmp_set_tap_count(unsigned char min_taps);
 int dmp_set_tap_time(unsigned short time);
 int dmp_set_tap_time_multi(unsigned short time);
-int dmp_set_shake_reject_thresh(long sf, unsigned short thresh);
+int dmp_set_shake_reject_thresh(int32_t sf, unsigned short thresh);
 int dmp_set_shake_reject_time(unsigned short time);
 int dmp_set_shake_reject_timeout(unsigned short time);
 
@@ -83,10 +83,10 @@ int dmp_enable_lp_quat(unsigned char enable);
 int dmp_enable_6x_lp_quat(unsigned char enable);
 
 /* Pedometer functions. */
-int dmp_get_pedometer_step_count(unsigned long *count);
-int dmp_set_pedometer_step_count(unsigned long count);
-int dmp_get_pedometer_walk_time(unsigned long *time);
-int dmp_set_pedometer_walk_time(unsigned long time);
+int dmp_get_pedometer_step_count(uint32_t *count);
+int dmp_set_pedometer_step_count(uint32_t count);
+int dmp_get_pedometer_walk_time(uint32_t *time);
+int dmp_set_pedometer_walk_time(uint32_t time);
 
 /* DMP gyro calibration functions. */
 int dmp_enable_gyro_cal(unsigned char enable);
@@ -94,8 +94,8 @@ int dmp_enable_gyro_cal(unsigned char enable);
 /* Read function. This function should be called whenever the MPU interrupt is
  * detected.
  */
-int dmp_read_fifo(short *gyro, short *accel, long *quat,
-    unsigned long *timestamp, short *sensors, unsigned char *more);
+int dmp_read_fifo(short *gyro, short *accel, int32_t *quat,
+    uint32_t *timestamp, short *sensors, unsigned char *more);
 
 #ifdef __cplusplus
 }
